// src/Types/generated/contentful.d.ts
export interface IAuthorFields {
  /** avatar */
  avatar?: Asset | undefined

  /** slug */
  slug?: string | undefined

  /** name */
  name?: string | undefined

  /** Credentials */
  credentials?: string | undefined
}

/** Author card for coaching */

export interface IAuthor extends Entry<IAuthorFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'author'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}
// ...
export interface ICoachingWeekFields {
  /** Cover photo */
  coverPhoto: Asset

  /** order */
  order: number

  /** Week Name */
  weekName?: string | undefined

  /** slug */
  slug: string

  /** Intro */
  intro?: string | undefined

  /** Description */
  weekDescription: Document

  /** Task count */
  taskCount?: number | undefined

  /** Lessons */
  lessons?: ILesson[] | undefined

  /** Locked */
  locked: boolean

  /** Week duration */
  duration: number
}

export interface ICoachingWeek extends Entry<ICoachingWeekFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'coachingWeek'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IHabitFields {
  /** title */
  title: string

  /** slug */
  slug: string

  /** period */
  period: 'MORNING' | 'EVENING' | 'AFTERNOON'

  /** description */
  description: Document
}

/** Example habits for user */

export interface IHabit extends Entry<IHabitFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'habit'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface ILessonFields {
  /** Lesson name */
  lessonName: string

  /** slug */
  slug: string

  /** Keywords */
  keywords?: string[] | undefined

  /** Author */
  author?:
    | 'Pietari Nurmi'
    | 'Eeva Siika-aho'
    | 'Anu-Katriina Pesonen'
    | 'Perttu Lähteenlahti'
    | 'Liisa Kuula-Paavola'
    | undefined

  /** Author card */
  authorCard?: IAuthor[] | undefined

  /** Lesson content */
  lessonContent?: Document | undefined

  /** Additional Information */
  additionalInformation?: Document | undefined

  /** Custom complete */
  customComplete?: string | undefined

  /** Stage */
  stage?: number | undefined

  /** Chronotype */
  chronotype?: 'MORNING_LARK' | 'NIGHT_OWL' | undefined

  /** cover */
  cover?: Asset | undefined

  /** section */
  section: ISection

  /** habit */
  habit?: IHabit[] | undefined

  /** weights */
  weights?: Record<string, any> | undefined
}

export interface ILesson extends Entry<ILessonFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'lesson'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}
// ...
export interface ISectionFields {
  /** order */
  order?: number | undefined

  /** title */
  title?: string | undefined

  /** description */
  description?: Document | undefined
}

export interface ISection extends Entry<ISectionFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'section'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}